Separate proxies from actions
