+++ pinax/stripe/apps.py
@@ -1,3 +1,5 @@
+import importlib
+
 from django.apps import AppConfig as BaseAppConfig
 from django.utils.translation import ugettext_lazy as _
 
@@ -7,3 +9,6 @@ class AppConfig(BaseAppConfig):
     name = "pinax.stripe"
     label = "pinax_stripe"
     verbose_name = _("Pinax Stripe")
+
+    def ready(self):
+        importlib.import_module("pinax.stripe.receivers")
