+++ pinax/stripe/management/commands/init_customers.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 
 from django.contrib.auth import get_user_model
 
-from ...actions import CustomerProxy
+from ...actions import customers
 
 
 class Command(BaseCommand):
@@ -12,5 +12,5 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         User = get_user_model()
         for user in User.objects.filter(customer__isnull=True):
-            CustomerProxy.create(user=user)
+            customers.create_customer(user=user)
             print("Created customer for {0}".format(user.email))
