+++ pinax/stripe/proxies/__init__.py
@@ -1,0 +1,7 @@
+from .charges import ChargeProxy  # noqa
+from .customers import CustomerProxy  # noqa
+from .events import EventProxy  # noqa
+from .exceptions import EventProcessingExceptionProxy  # noqa
+from .invoices import InvoiceProxy  # noqa
+from .subscriptions import CurrentSubscriptionProxy  # noqa
+from .transfers import TransferProxy  # noqa
