+++ pinax/stripe/proxies/invoices.py
@@ -1,0 +1,22 @@
+import stripe
+
+from .. import models
+
+
+class InvoiceProxy(models.Invoice):
+
+    class Meta:
+        proxy = True
+        ordering = ["-date"]
+
+    def retry(self):
+        if not self.paid and not self.closed:
+            inv = stripe.Invoice.retrieve(self.stripe_id)
+            inv.pay()
+            return True
+        return False
+
+    def status(self):
+        if self.paid:
+            return "Paid"
+        return "Open"
