+++ pinax/stripe/signals.py
@@ -1,8 +1,13 @@
 from django.dispatch import Signal
 
 
+charge_refunded = Signal(providing_args=["charge_proxy"])
+charge_captured = Signal(providing_args=["charge_proxy"])
 cancelled = Signal(providing_args=["stripe_response"])
 card_changed = Signal(providing_args=["stripe_response"])
+charge_event_received = Signal(providing_args=["event"])
+customer_subscription_event = Signal(providing_args=["event", "customer"])
+invoice_event_received = Signal(providing_args=["event"])
 subscription_made = Signal(providing_args=["plan", "stripe_response"])
 webhook_processing_error = Signal(providing_args=["data", "exception"])
